Refactor import paths and enhance logging in hyper_search
